Opnieuw minder gevallen waarin de container niet lager geplaatst kan worden + extra constraint

+++ src/main/java/Coördinaat.java
@@ -22,7 +22,7 @@ public class Coördinaat {
         this.x = x;
     }
 
-    public float getY() {
+    public int getY() {
         return y;
     }
